backend/pyproject.toml
[project]
name = "songhut-backend"
version = "2.0.0"
description = "SongHut 2.0 Backend - FastAPI + PostgreSQL + Celery"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115,<0.116",
    "uvicorn[standard]>=0.32",
    "sqlalchemy[asyncio]>=2.0,<2.1",
    "asyncpg>=0.29",
    "alembic>=1.14",
    "pydantic>=2.9",
    "pydantic-settings>=2.6",
    "python-jose[cryptography]>=3.3",
    "passlib[bcrypt]>=1.7",
    "python-multipart>=0.0.16",
    "aiofiles>=24.1",
    "celery[redis]>=5.4",
    "redis>=5.2",
    "structlog>=24.4",
    "httpx>=0.28",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3",
    "pytest-asyncio>=0.24",
    "httpx>=0.28",
    "ruff>=0.8",
    "mypy>=1.13",
]
algorithms = [
    "numpy>=1.26",
    "librosa>=0.10",
    "pretty-midi>=0.3",
    "music21>=9.3",
    "scipy>=1.14",
]

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B", "SIM"]

[tool.mypy]
python_version = "3.12"
strict = false
ignore_missing_imports = true

[build-system]
requires = ["setuptools>=69"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
